pyproject.toml
[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
natricine = { workspace = true }
natricine-redisstream = { workspace = true }
natricine-conformance = { workspace = true }
natricine-aws = { workspace = true }
natricine-otel = { workspace = true }
natricine-sql = { workspace = true }
natricine-http = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-anyio>=0.0.0",
    "ruff>=0.8",
    "pydantic>=2.0",
    "ty>=0.0.6",
    # Backend implementations for testing
    "natricine-redisstream",
    "natricine-conformance",
    # AWS type stubs for typechecking
    "natricine-aws[types,test]",
    # OTEL for testing
    "natricine-otel[test]",
    # Redis Streams with testcontainers
    "natricine-redisstream[test]",
    # SQL with SQLite for testing
    "natricine-sql[test]",
    # HTTP for testing
    "natricine-http[test]",
    # Documentation
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-llmstxt>=0.5",
    "testcontainers>=4.13.3",
    "mkdocs-macros-plugin>=1.5.0",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "PLC0415", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["natricine"]

[tool.pytest.ini_options]
testpaths = ["tests", "packages/*/tests"]
markers = [
    "containers: tests that require Docker/testcontainers",
    "interop: tests for Go/Python wire format compatibility",
]

[tool.ty.environment]
extra-paths = [
    "packages/natricine/src",
    "packages/natricine-redisstream/src",
    "packages/natricine-conformance/src",
    "packages/natricine-aws/src",
    "packages/natricine-otel/src",
    "packages/natricine-sql/src",
    "packages/natricine-http/src",
]
